requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
numpy==1.26.1
pandas==2.1.2
openpyxl==3.1.2
plotly==5.18.0
kaleido==0.2.1
